Implemented RecentUnlocks and created most of the  model classes thats needed for it, still lacks the Status model class.

+++ WaniKaniClient/Models/StudyQueue.cs
@@ -15,7 +15,7 @@ namespace WaniKaniClient.Models
         public int ReviewsAvailable { get; set; }
 
         [JsonProperty("next_review_date")]
-        public int NextReviewTimeStamp { get; set; }
+        public long NextReviewTimeStamp { get; set; }
         public DateTime NextReviewDate
         {
             get
